works on theme change
